+++ CPE200Lab1/CPE200Lab1/CalculatorEngine.cs
@@ -26,30 +26,7 @@ namespace CPE200Lab1
             return false;
         }
 
-        public string Process(string str)
-        {
-            //Split input string to multiple parts by space
-            List<string> parts = str.Split(' ').ToList<string>();
-            string result;
-            //As long as we have more than one part
-            while(parts.Count > 1)
-            {
-                //Check if the first three is ready for calcuation
-                if(!(isNumber(parts[0]) && isOperator(parts[1]) && isNumber(parts[2])))
-                {
-                    return "E";
-                } else
-                {
-                    //Calculate the first three
-                    result = calculate(parts[1], parts[0], parts[2], 4);
-                    //Remove the first three
-                    parts.RemoveRange(0, 3);
-                    // Put back the result
-                    parts.Insert(0, result);
-                }
-            }
-            return parts[0];
-        }
+        
         public string unaryCalculate(string operate, string operand, int maxOutputSize = 8)
         {
             switch (operate)
