+++ CPE200Lab1/CPE200Lab1/MainForm.cs
@@ -47,7 +47,7 @@ namespace CPE200Lab1
         {
             InitializeComponent();
             memory = 0;
-            engine = new CalculatorEngine();
+            engine = new SimpleCalculatorEngine();
             resetAll();            
         }
         private void btnSwitch_Click(object sender, EventArgs e)
